move pbf to subdir
